SN1-138: Removed unnecessary files and renamed/moved others
