+++ pyproject.toml
@@ -22,6 +22,7 @@ exclude = '''
 | build
 )/
 '''
+
 [tool.isort]
 atomic = true
 profile = "black"
